pyproject.toml
[project]
name = "agent-debate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "autogen>=0.11.1",
    "openai>=2.21.0",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.2.1",
]
